# docs/openapi/cxp-platform-api-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CXP Platform API",
    "description": "Civic Exchange Platform — citizen service-request intake, AI enrichment, governance discovery, and announcement feed. Authenticated endpoints require a Bearer JWT (Platform HS256 for operators, Supabase ES256 for citizens). Public endpoints are explicitly marked.\n\n## Compatibility & Deprecation Policy\n\nThe platform maintains backward-compatible aliases for one release cycle while clients migrate to canonical field names. All deprecated fields and endpoints are marked with a strikethrough in Swagger UI and carry `deprecated: true` in the generated OpenAPI schema — SDK generators will surface these as deprecated in generated client code.\n\n**Active transitional aliases (marked deprecated in schema):**\n- `ServiceRequestEvent.oldValue` → migrate to `oldStatus`\n- `ServiceRequestEvent.newValue` → migrate to `newStatus`\n- `ServiceRequestEvent.changedByRole` → migrate to `actorType`\n- `CitizenProfile.appUserId` → migrate to `id`\n- `CitizenProfile.postalCode` → migrate to `pinCode`\n- `UserProvision.appUserId` → migrate to `id`\n- `ServiceRequestCreate.aiConversationId` → not replaced; field is legacy from a discontinued conversational flow\n- Pagination `number` field → migrate to `page`\n\n**Deprecated endpoints:**\n- `POST /api/v1/ai/service-requests/analyze` → migrate to `POST /api/v1/service-requests/analyze` (identical contract)",
    "version": "v1"
  },
  "servers": [
    {
...
    }
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/announcements": {
      "get": {
        "tags": [
          "Announcements"
        ],
        "summary": "List all announcements",
        "description": "Returns all announcements for the operator's tenant, including drafts and expired entries. Operator-only — citizens use /feed or /my-feed instead.",
        "operationId": "list_1",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "403": {
            "description": "Forbidden — the caller does not have the OPERATOR role."
          },
          "200": {
            "description": "All announcements for the operator's tenant",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Announcements"
        ],
        "summary": "Create an announcement",
        "description": "Publishes a civic announcement scoped to the operator's tenant. Targeting (ward, city, district, or state) is set in the request body. Operator-only — requires the OPERATOR role on the platform JWT.",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure — required fields missing, invalid targeting, or targeting conflicts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "201": {
            "description": "Announcement created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "403": {
            "description": "Forbidden — the caller does not have the OPERATOR role."
          }
        }
      }
    },
...
    "/api/v1/service-requests/{id}/status": {
      "patch": {
        "tags": [
          "Service Requests"
        ],
        "summary": "Update service request status",
        "description": "Advances the lifecycle state of a service request. Operator-only — requires the OPERATOR role on the platform JWT. Each transition is recorded as an immutable event in the request's event log.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequestStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Validation failure or invalid status transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "Service request status updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — the caller does not have the OPERATOR role."
          },
          "404": {
            "description": "Service request not found for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/announcements/{id}": {
      "get": {
        "tags": [
          "Announcements"
        ],
        "summary": "Retrieve an announcement",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Announcement not found for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Announcement returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          }
        }
      }
    },
    "/api/v1/announcements/my-feed": {
      "get": {
        "tags": [
          "Announcements"
        ],
        "summary": "Retrieve the personalised citizen announcement feed",
        "description": "Returns published, active, non-expired announcements relevant to the authenticated citizen. Geography (ward, city, district, state) is resolved from the caller's stored profile — no geography parameters are accepted. Requires: authenticated citizen session.",
        "operationId": "myFeed",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected platform error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required — Bearer JWT is missing, expired, or invalid. Response body: {\"error\":\"Unauthorized\",\"message\":\"Invalid or missing token\"}"
          },
          "200": {
            "description": "Personalised announcement feed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CitizenFeedResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/announcements/feed": {
      "get": {
        "tags": [
          "Announcements"
        ],
        "summary": "Browse the public announcement feed",
        "description": "Returns published, active, non-expired announcements. No authentication required. Optionally filter by wardId to narrow to a specific ward's feed. Authenticated citizens with a stored profile should use /my-feed for personalised results.",
        "operationId": "feed",
        "parameters": [
          {
            "name": "wardId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Public announcement feed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementResponse"
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "UTC timestamp of the error",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Machine-readable error code, e.g. VALIDATION_FAILED, NOT_FOUND, BAD_REQUEST",
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "path": {
            "type": "string",
            "description": "Request path that produced the error",
            "example": "/api/v1/service-requests"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "Field-level validation errors; present only on VALIDATION_FAILED responses",
              "nullable": true
            },
            "description": "Field-level validation errors; present only on VALIDATION_FAILED responses",
            "nullable": true
          }
        },
        "description": "Standard platform error response. The 'fields' map is present only on VALIDATION_FAILED (400) responses — it maps each failing field name to its constraint message. All other error types omit 'fields'."
      },
...
      "AnnouncementCreateRequest": {
        "required": [
          "content",
          "targetScope",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string",
            "description": "Announcement headline. Required."
          },
          "content": {
            "type": "string",
            "description": "Full announcement body text. Required."
          },
          "summary": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Short summary for list views and notifications. Optional.",
            "nullable": true
          },
          "category": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "description": "Operator-defined category label (e.g. 'infrastructure', 'health'). Optional.",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "description": "Priority level. Optional — defaults to NORMAL.",
            "nullable": true,
            "enum": [
              "NORMAL",
              "IMPORTANT",
              "URGENT",
              "CRITICAL"
            ]
          },
          "status": {
            "type": "string",
            "description": "Initial publication status. Optional — defaults to DRAFT. ARCHIVED and EXPIRED are system-managed and rejected on creation.",
            "nullable": true,
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED",
              "EXPIRED"
            ]
          },
          "targetScope": {
            "type": "string",
            "description": "Geographic scope of this announcement. Required. Determines which citizens receive the announcement in their feed.",
            "enum": [
              "WARD",
              "ZONE",
              "CITY",
              "DISTRICT",
              "STATE"
            ]
          },
          "wardId": {
            "type": "string",
            "description": "Ward identifier. Required when targetScope is WARD; null otherwise.",
            "format": "uuid",
            "nullable": true
          },
          "zoneId": {
            "type": "string",
            "description": "Zone identifier. Required when targetScope is ZONE; null otherwise.",
            "format": "uuid",
            "nullable": true
          },
          "cityId": {
            "type": "string",
            "description": "City identifier. Required when targetScope is CITY; null otherwise.",
            "format": "uuid",
            "nullable": true
          },
          "districtId": {
            "type": "string",
            "description": "District identifier. Required when targetScope is DISTRICT; null otherwise.",
            "format": "uuid",
            "nullable": true
          },
          "stateId": {
            "type": "string",
            "description": "State identifier. Required when targetScope is STATE; null otherwise.",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "description": "UTC expiry timestamp. Null means the announcement never auto-expires.",
            "format": "date-time",
            "nullable": true
          },
          "pinned": {
            "type": "boolean",
            "description": "When true, the announcement is pinned to the top of citizen feeds. Optional — defaults to false.",
            "nullable": true
          },
          "tags": {
            "type": "string",
            "description": "Topic labels as a JSON array string (e.g. '[\"water\",\"road\"]'). Optional — stored verbatim.",
            "nullable": true
          }
        },
        "description": "Operator request to create a new civic announcement. Exactly one geographic FK field must be provided to match the targetScope. The service validates this invariant and returns HTTP 400 on violation."
      },
      "AnnouncementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable platform-generated identifier for this announcement.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Announcement headline."
          },
          "content": {
            "type": "string",
            "description": "Full announcement body text."
          },
          "summary": {
            "type": "string",
            "description": "Short summary for list views and notifications.",
            "nullable": true
          },
          "category": {
            "type": "string",
            "description": "Operator-defined category label.",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "description": "Priority level.",
            "enum": [
              "NORMAL",
              "IMPORTANT",
              "URGENT",
              "CRITICAL"
            ]
          },
          "status": {
            "type": "string",
            "description": "Current publication status. ARCHIVED and EXPIRED are system-managed transitions.",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED",
              "EXPIRED"
            ]
          },
          "governingBodyId": {
            "type": "string",
            "description": "Canonical governing authority that owns this announcement.",
            "format": "uuid"
          },
          "targetScope": {
            "type": "string",
            "description": "Geographic scope determining which citizens receive this announcement.",
            "enum": [
              "WARD",
              "ZONE",
              "CITY",
              "DISTRICT",
              "STATE"
            ]
          },
          "wardId": {
            "type": "string",
            "description": "Ward identifier. Non-null only when targetScope is WARD.",
            "format": "uuid",
            "nullable": true
          },
          "zoneId": {
            "type": "string",
            "description": "Zone identifier. Non-null only when targetScope is ZONE.",
            "format": "uuid",
            "nullable": true
          },
          "cityId": {
            "type": "string",
            "description": "City identifier. Non-null only when targetScope is CITY.",
            "format": "uuid",
            "nullable": true
          },
          "districtId": {
            "type": "string",
            "description": "District identifier. Non-null only when targetScope is DISTRICT.",
            "format": "uuid",
            "nullable": true
          },
          "stateId": {
            "type": "string",
            "description": "State identifier. Non-null only when targetScope is STATE.",
            "format": "uuid",
            "nullable": true
          },
          "publishedAt": {
            "type": "string",
            "description": "UTC timestamp when the announcement was published. Null when status is DRAFT.",
            "format": "date-time",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "description": "UTC timestamp when the announcement auto-expires. Null means no automatic expiry.",
            "format": "date-time",
            "nullable": true
          },
          "pinned": {
            "type": "boolean",
            "description": "When true, the announcement is pinned to the top of citizen feeds."
          },
          "tags": {
            "type": "string",
            "description": "Topic labels stored as a JSON array string (e.g. '[\"water\",\"road\"]'). Null when no tags were supplied.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "UTC creation timestamp.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "UTC last-updated timestamp.",
            "format": "date-time"
          }
        },
        "description": "Full operator-facing announcement representation. Includes internal routing fields (geographic FKs, governingBodyId, status) not exposed in the citizen feed view."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Bearer JWT. Accepted formats: Platform HS256 (operator — carries governingBodyId claim) or Supabase ES256 (citizen — no tenantId claim). Header: Authorization: Bearer <token>",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}